박스 맵 파일 세이브 로드, 메인 게임에서 로드

+++ 숙제/3개월차/팀 과제(2차-록맨)/8일차-라인 에디터, 저장, 불러오기, 스크롤/DefaultWindow/DefaultWindow/CMainGame.h
@@ -21,7 +21,7 @@ public:
 private:
 	HDC			m_hDC;
 
-	DWORD		m_ullTime;
+	ULONGLONG		m_ullTime;
 	int			m_iFPS;
 	TCHAR		m_szFPS[16];
